finished back propagation

+++ src/main/scala/models/Inputs.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Inputs(values: Vector[Double], expectedClass: Option[Int])
+case class Inputs(values: Vector[Double], expectedClass: Option[Double])
